# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/scholarships": {
      "get": {
        "operationId": "ScholarshipsController_getScholarships",
        "parameters": [
          {
            "name": "studyLevel",
            "required": false,
            "in": "query",
            "description": "Filter by study level: foundation, diploma, degree",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by scholarship type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "description": "Filter by location: local, overseas, both",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fieldId",
            "required": false,
            "in": "query",
            "description": "Filter by field ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "all",
            "required": false,
            "in": "query",
            "description": "Include all statuses (for admin)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scholarships retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all scholarships",
        "tags": [
          "Scholarships"
        ]
      },
      "post": {
        "operationId": "ScholarshipsController_createScholarship",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScholarshipRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scholarship created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create a new scholarship (admin only)",
        "tags": [
          "Scholarships"
        ]
      }
    },
    "/api/scholarships/{id}": {
      "get": {
        "operationId": "ScholarshipsController_getScholarshipById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Scholarship ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scholarship retrieved successfully"
          },
          "404": {
            "description": "Scholarship not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get scholarship by ID",
        "tags": [
          "Scholarships"
        ]
      },
      "put": {
        "operationId": "ScholarshipsController_updateScholarship",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Scholarship ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScholarshipRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scholarship updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Scholarship not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Update an existing scholarship (admin only)",
        "tags": [
          "Scholarships"
        ]
      },
      "delete": {
        "operationId": "ScholarshipsController_deleteScholarship",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Scholarship ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scholarship deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Scholarship not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Delete a scholarship (admin only - permanent delete)",
        "tags": [
          "Scholarships"
        ]
      }
    },
    "/api/scholarships/level/{level}": {
      "get": {
        "operationId": "ScholarshipsController_getScholarshipsByLevel",
        "parameters": [
          {
            "name": "level",
            "required": true,
            "in": "path",
            "description": "Study level: foundation, diploma, or degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scholarships retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get scholarships by study level",
        "tags": [
          "Scholarships"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your Supabase access token"
      }
    },
    "schemas": {
...
      "CreateScholarshipRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Scholarship name"
          },
          "organization_name": {
            "type": "object",
            "description": "Organization name"
          },
          "amount": {
            "type": "number",
            "description": "Scholarship amount"
          },
          "type": {
            "type": "string",
            "description": "Scholarship type",
            "enum": [
              "Merit-based",
              "Need-based",
              "Academic",
              "Other"
            ]
          },
          "location": {
            "type": "object",
            "description": "Location"
          },
          "deadline": {
            "type": "object",
            "description": "Application deadline"
          },
          "description": {
            "type": "object",
            "description": "Description"
          },
          "level": {
            "type": "object",
            "description": "Level (study level)"
          },
          "field_id": {
            "type": "number",
            "description": "Field ID"
          },
          "website_url": {
            "type": "object",
            "description": "Website URL"
          },
          "contact_email": {
            "type": "object",
            "description": "Contact email"
          },
          "contact_phone": {
            "type": "object",
            "description": "Contact phone"
          },
          "requirements": {
            "type": "object",
            "description": "Requirements (text)"
          },
          "benefits": {
            "type": "object",
            "description": "Benefits (text)"
          },
          "success_rate": {
            "type": "number",
            "description": "Success rate"
          },
          "status": {
            "type": "string",
            "description": "Scholarship status",
            "enum": [
              "active",
              "expired",
              "draft"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateScholarshipRequestDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}